don't replace /chan, don't crash when button is released outside qwm's screen.
by not replacing /chan, we let earlier binds on chan, e.g. by #^
for the snarf device, be used by programs in qwm.  this unbreaks
copy/paste.

a button can be pressed inside qwm and released outside.  this
caused qwm not to find the column/window where the button was
released, and dereference a nil.