give ptrs to the right window when holding button down.  fix an obscure crash.  don't forkns for "^q c".
when pressing a button in one window and moving the pointer over
another, that other window got the focus.  this is not uncommon,
e.g. when dragging a scrollbar, adjacent to another window.  we now
give ptrs to the areas/windows where a button was first pressed.
only on release of the last button do we consider changing focus.

the obscure crash was to release the button over the column tag bar
when dragging a window by its title bar.

forking the namespace for "^q c" isn't necessary, only makes it
impossible to use a command that changes qwm's namespace.  not
normally advised, but no reason to limit.  fwiw, wm/sh and wm/run
both fork the namespace.