implementing multiple windows (in a draw screen) for a client.  makes menu(9)'s work properly.
for tk menu(9) widgets, tk asks for a new window to draw the widget
on.  it does so by sending a !reshape request with a tag other than
".".  before this, we just returned the normal program's widget.
now we create a new window in the screen for this.  we assume the
client keeps a reference to the image, so it'll go away automatically
when the client drops the last reference to it and we don't have
to do anything for the "delete" request.

we also have Refbackup for all images now.  obviously necessary
since we draw menu's over our normal windows.