don't resize as often, and improve Mstack vs Msingle handling.
we don't keep a "layout rect" (for Mstack) for each window anymore,
we used to keep that even for Msingle columns.  we only have "want
rect" now, and it is set based on the column mode.  when we switch
modes, we set new want rects.  for Msingle, windows are always
the full column size.  the focused window is on top (screen-wise).
this has a down side:  lots of memory consumption.  perhaps i'll
give up Refbackup for windows, and force programs to redraw.  will
need handling of apps that create additional windows (eg menu(9)).

we resize less because we simply move windows in hidden columns
off-screen, by changing the window origins.  for stack mode, we
also just change origin if the window is moved but not resized.

also add shortcuts ^qd for toggling debug printing.
and ^qD for dumping the window state.