don't use Refbackup for windows
we nog longer create windows (in our screen) with Refbackup set.
instead, we use Refnone.  so we don't have a backup of images.

not having an image backup is not too annoying because qwm is tiling.
we simply ask applications to redraw to their image (we have to
give them a new image for this though...).  we still move hidden
windows off-screen, so those applications don't have to redraw
(until they are visible again).  windows in stack mode only need
to redraw when their dx or dy change, not when their position
changes.  windows in single mode have to redraw when they become
visible.

windows also have to redraw when another window was shown over it,
eg a menu(9).

it seems to work reasonably well.  this also reduces image memory
consumption, and lets you have many more windows alive than before
(or than with wm/wm).