window placement has changed a lot, and more.
we show two columsn by default now.  col 2 on the right, smaller
than col 1, but big enough so all initial windows (from the profile)
go there.

add key 'r', for unwarping the mouse.
it replaces 'n' (go to "other" window).
and 'm' is gone too, was similar to 'n'.

new windows go to an adjacent column if it is >=1.5 times as large
as the current column.

new windows take the place of the biggest column (focused window
ignored).