give windows adjusted ptrs when their origin has changed.
before this, a moved (with same size) window in stacked mode would
get mouse events for some offset of the real coordinates.  it also
caused clicks on the titlebar to be ignored.

this can probably be done better, but the semantics of origin() for
screen images still isn't completely clear to me.