on "delete" of a window (eg menu), force redraws for all windows that were affected.
a new window (eg menu) can be drawn over adjacent windows.  simple
to reproduce in wm/sh and wm/run by pressing b2 near the edge of a
window, with another window close by.  when that window was deleted,
the other app didn't redraw, so part of the menu remained visible.